Root파일명 App으로 변경, ThemeProvider이하 요소 Root(App)로 이동
