--- src/Root.tsx
@@ -1,11 +1,0 @@
-import Coins from "./routes/Coins";
-
-function Root() {
-  return (
-    <>
-      <Coins />
-    </>
-  );
-}
-
-export default Root;
